Add clock functionality
You can now clock in to single tasks. The task will get a @time
attribute to indicate the time already spend on the task in
the format d.hh.mm and a @clock attribute with a time stamp
to calculate the new time value. You will be clocked out
if the task is done or canceled.